sweetalert and add task view

+++ app/Http/Controllers/UsersController.php
@@ -8,18 +8,28 @@ use Illuminate\Auth\Access\Gate;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Validator;
+use RealRashid\SweetAlert\Facades\Alert;
 
 class UsersController extends Controller
 {
 
 ####################### view All Users #########################
     public function index(){
+        abort_unless( \Gate::allows('users-access') ,403);
+
+        //if session key not define in ToSweetAlert file
+        if(session('success_message')){
+            Alert::success('Success!',session('success_message'));
+//            Alert::image('User image!','User profile image', asset('dist/img/user3-128x128.jpg'));
+        }
+
         $users = User::all();
         return view('Users.index',compact('users'));
     }
 ####################### delete User #########################
     public function destroy($id)
     {
+        abort_unless( \Gate::allows('users-delete') ,403);
         $user = User::findOrFail($id);
         $user->delete();
         //Redirect to a specified route with flash message.
@@ -30,6 +40,8 @@ class UsersController extends Controller
 ####################### edit User #########################
 
     public function edit($id){
+        abort_unless( \Gate::allows('users-edit') ,403);
+
         $user = User::findOrFail($id);
         $roles = Role::all();
         return view('users.edit',compact(['user','roles']));
@@ -37,8 +49,10 @@ class UsersController extends Controller
 ## update and Assign Roles and Edit Telegram settings for All User ##
 
     public function update(Request $request,User $user){
+        abort_unless( \Gate::allows('users-update') ,403);
 
         $data = $this->validator($request->all())->validate();
+
         $user->roles()->sync($request->roles);
         $user->name = $data['name'];
         $user->email = $data['email'];
@@ -47,10 +61,17 @@ class UsersController extends Controller
             $password = Hash::make($password_valid['password']);
             $user->password = $password; //hashed password.
         }
-       if($user->save())
-            $request->session()->flash('success','تم تحديث المستخدم بنجاح');
-        else
-            $request->session()->flash('error','يوجد مشكلة في البيانات المدخلة');
+       if($user->save()){
+//           $request->session()->flash('success','تم تحديث المستخدم بنجاح');
+//           Alert::success('Success User', 'تم تحديث المستخدم بنجاح');
+//           Alert::Image('User image!','User profile image', asset('dist/img/user1-128x128.jpg'));
+           Alert::toast('Toast Message', 'info');
+
+       }
+        else{
+            $request->session()->flash('errors','يوجد مشكلة في البيانات المدخلة');
+        }
+
         $users = User::all();
         return redirect('/users');
     }
